Move email logic to logic.email package
