+++ src/test/java/teammates/logic/core/FeedbackSessionsLogicTest.java
@@ -28,6 +28,7 @@ import teammates.common.exception.InvalidFeedbackSessionStateException;
 import teammates.common.exception.InvalidParametersException;
 import teammates.common.util.Const;
 import teammates.common.util.TimeHelper;
+import teammates.logic.email.FeedbackSessionsEmailsLogic;
 import teammates.storage.api.FeedbackSessionsDb;
 import teammates.storage.entity.Course;
 import teammates.storage.entity.FeedbackQuestion;
