+++ src/test/java/teammates/logic/email/DeadlineExtensionsEmailsLogicTest.java
@@ -1,4 +1,4 @@
-package teammates.logic.core;
+package teammates.logic.email;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -14,7 +14,6 @@ import teammates.common.util.EmailType;
 import teammates.common.util.EmailWrapper;
 import teammates.common.util.TaskWrapper;
 import teammates.logic.api.MockTaskQueuer;
-import teammates.logic.email.EmailQueueService;
 import teammates.logic.email.model.DeadlineExtensionUpdateEmailContext;
 import teammates.logic.email.model.EmailContact;
 import teammates.logic.email.model.FeedbackSessionEmailContext;
